% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptoticDecayModel.R
\name{asymptoticDecayFit}
\alias{asymptoticDecayFit}
\title{Fit an asymptotic decay model to reach deviations.}
\usage{
asymptoticDecayFit(schedule, signal, gridpoints = 11, gridfits = 10)
}
\arguments{
\item{schedule}{A vector of length N with the perturbation schedule.}

\item{reaches}{A vector of length N with reach deviation data.}
}
\value{
A named numeric vector with the optimal parameter that fits a simple
rate model to the data as best as possible, with these elements:
- lambda: the rate of change
- N0: the asymptote
}
\description{
This function is part of a set of functions to fit and 
evaluate a simple learning rate model of motor learning.

The schedule should usually be a sequence of ones. The reach deviations have 
to be baselined (but the baseline is cut from the data).
}
\details{
?
}
\examples{
# write example!
}
