% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptoticDecayModel.R
\name{asymptoticDecayModel}
\alias{asymptoticDecayModel}
\title{Execute a simple model given parameters and a reach 
deviation schedule.}
\usage{
asymptoticDecayModel(par, schedule)
}
\arguments{
\item{par}{A named vector with the model parameter (see details).}

\item{schedule}{A vector of length N with the perturbation schedule.}
}
\value{
A data frame with one column: `output`, and N rows, so that each row
has the output of the modeled process on each trials.
}
\description{
This function is part of a set of functions to fit and 
evaluate an exponential decay model with asymptote.
}
\details{
The `par` argument is a named numeric vector that should have the
following element:
- lambda: learning rate
- N0: asymptote

The schedule usually consists of a sequence of ones. It will be multiplied
by the asymptote.
}
\examples{
?
}
