% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptoticDecayModel.R
\name{asymptoticDecayMSE}
\alias{asymptoticDecayMSE}
\title{Get the MSE for how well an asymptotic decay model fits reaches.}
\usage{
asymptoticDecayMSE(par, schedule, signal)
}
\arguments{
\item{par}{A named numeric vector with the model parameter (see 
asymptoticDecayModel).}

\item{schedule}{A numeric vector of length N with the perturbation schedule.}

\item{reaches}{A numeric vector of length N with reach deviations matching
the perturbation schedule.}
}
\value{
A float: the mean squared error between the total model output and
the reach deviations.
}
\description{
This function is part of a set of functions to fit and 
evaluate exponential decay model with asymptote..
}
\details{
The `par` argument is a named numeric vector that should have the
following element:
- lambda: the learning rate
- N0: the asymptote

The schedule is usually a sequence of ones, which is multiplied by the 
asymptote in the function.
}
\examples{
?
}
