% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalFit.R
\name{bootstrapNormalPeakCI}
\alias{bootstrapNormalPeakCI}
\title{Bootstrap the 95\% confidence interval of a normal function.}
\usage{
bootstrapNormalPeakCI(pos, values, bootstraps = 1000)
}
\arguments{
\item{pos}{A numeric vector with N data positions.}

\item{values}{A numeric matrix with N rows and M cols giving values 
(densities) from M observations at N positions.}

\item{bootstraps}{The number of bootstrapped samples (M columns) to fit
a normal function to.}
}
\value{
Named numeric vector, with the 2.5^nd, 50^th and 97.5^th percentile 
of the locations of peaks of a normal function fitted to samples of the 
data. These are the upper and lower bound of the 95% confidence interval as
well as the median.
}
\description{
?
}
\details{
?
}
\examples{
?
}
