% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrationPolynomial}
\alias{calibrationPolynomial}
\title{Apply calibration polynomial to new measurements.}
\usage{
calibrationPolynomial(P, Mcoords)
}
\arguments{
\item{P}{Vector of parameters fit by calibrationParameters() on other data.}

\item{Mcoords}{A data frame with measured coordinates in columns x and y.}
}
\value{
A data frame with estimated coordinates in target reference frame.
}
\description{
Not yet.
}
\details{
Not yet.
}
\examples{
data("localization")
str(localization_aligned)

# let's see if we can convert the touch screen pixels to (centi)meters
train <- localization_aligned[c(1:16),]
test <- localization_aligned[c(17:24),]

# get parameters fit onto the training data:
trainTaps <- train[,which(names(train) \%in\% c('tapx_px', 'tapy_px'))]
names(trainTaps) <- c('x','y')
trainHand <- train[,which(names(train) \%in\% c('handx_m', 'handy_m'))]
names(trainHand) <- c('x','y')
params <- calibrationParameters(trainTaps,trainHand)

# get test data ready
testTaps <- test[,which(names(train) \%in\% c('tapx_px', 'tapy_px'))]
names(testTaps) <- c('x','y')

calibrationPolynomial(params,testTaps)
}
