% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleFit.R
\name{circleErrors}
\alias{circleErrors}
\title{Get mean squared error between coordinates and a circle}
\usage{
circleErrors(freepar, X, Y, setpar = c())
}
\arguments{
\item{freepar}{Vector with xc and yc parameters: x and y of the circle's middle.
Optionally also includes \code{r}: the radius of the circle.}

\item{X}{Vector of X coordinates}

\item{Y}{Vector of Y coordinates}

\item{setpar}{Vector that can have a fixed radius \code{r}, or can be empty.}
}
\value{
The mean squared error between the distances of \code{X} and 
\code{Y} from the position in par and the radius \code{r}.
}
\description{
Get mean squared error between coordinates and a circle
}
