% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleFit.R
\name{circleFit}
\alias{circleFit}
\title{Fit a circle to data}
\usage{
circleFit(X, Y, r = 1, fitr = FALSE)
}
\arguments{
\item{X}{Vector of X coordinates}

\item{Y}{Vector of Y coordinates}

\item{r}{Radius of the circle the coordinates should be on (default: 1).}

\item{fitr}{(boolean) Should radius be fit? (default: FALSE)}
}
\description{
Fit a circle to data
}
