% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceEllipse.R
\name{confidenceEllipse}
\alias{confidenceEllipse}
\title{Get coordinates of a confidence ellipse.}
\usage{
confidenceEllipse(x, y = NA, interval = 0.95, vectors = 100)
}
\arguments{
\item{interval}{The proportion of data that should fall in the confidence
ellipse. Default: 0.95.}

\item{vectors}{Number of points on the ellipse, equally spaced angularly,
not in absolute distance.}

\item{X}{Vector of X-coordinates or a Mx2 matrix with M sets of X and Y 
coordinates.}

\item{Y}{If X is not a matrix, but a vector this has to be a vector of 
equal size with Y coordinates.}
}
\value{
A data frame with `vectors` rows, and 2 columns: `x` and `y` with
coordinates of an ellipse containing `interval` proportion of the data.
}
\description{
?
}
\details{
?
}
\examples{
?
}
