% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert2cm}
\alias{convert2cm}
\title{Conversion from the standard units of our lab's equipment to centimeters}
\usage{
convert2cm(df, scale = NULL, home = NULL, from, replace = TRUE)
}
\arguments{
\item{df}{A data frame with columns that have to be converted.}

\item{scale}{Scale from (unit) to centimeters.}

\item{home}{Location of home position, which will be (0,0).}

\item{from}{The type of equipment the data is measured with.}

\item{replace}{Replace the old columns, or keep them.}
}
\value{
The data frame is returned, with the relevant columns converted.
}
\description{
Perhaps we don't need this function anymore.
}
\details{
?
}
\examples{
?
}
