% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pavlovia.R
\name{convertCellToNumVector}
\alias{convertCellToNumVector}
\title{Convert list in Pavlovia CSV to R vector}
\usage{
convertCellToNumVector(v)
}
\arguments{
\item{v}{Text version of Pavlovia list}
}
\value{
A numeric R vector
}
\description{
does not yet allow non-numeric data types
}
\details{
?
}
\examples{
?
}
