% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{coordW}
\alias{coordW}
\title{Get parameter weights for converting between reference frames.}
\usage{
coordW(Mcoords, Tcoords)
}
\arguments{
\item{Mcoords}{A data frame with x and y columns of measurements}

\item{Tcoords}{A vector with target coordinates along one axis.}
}
\value{
Parameter weights.
}
\description{
Adapted from some old Python code. Used by calibrate().
No need to understand or use this
}
\details{
Not yet.
}
\examples{
Not yet.
}
