% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics2d.R
\name{density2D}
\alias{density2D}
\title{Create matrix representing 2D density.}
\usage{
density2D(x, y, bw = 1, weights = NULL, n = 100, from = NULL,
  to = NULL, cut = 3, na.rm = FALSE)
}
\arguments{
\item{x}{Vector of X coordinates for density estimation.}

\item{y}{Vector of Y coordinates for density estimation of the same length
as `x`.}

\item{bw}{Width of the Gaussian kernel, can be a single number or have two
values, to be applied along x and y axes, respectively. In contrast to
`density()` there is no good default value (set to 1 for now).}

\item{n}{Number of points at which to get density estimates, can be a single
integer, or two, to be applied along the x and y axes respectively.}

\item{from}{Start of the interval in which to do density estimation, can be
a single value, or two, to be applied along the x and y axes respectively.}

\item{to}{End of the interval in which to do density estimation, can be a
single value, or two, to be applied along the x and y axes respectively.}

\item{cut}{Used if `from` and `to` are not given: density is estimated for
the minimum and maximum of the `x` of the `y` vectors plus/minus cut*bw.}

\item{na.rm}{Boolean: remove NA values or not.}
}
\value{
The functions returns a list with 3 entries:
    x: x coordinates of density estimates
    y: y coordinates of density estimates
    z: a matrix of size (x,y) with density estimates
These can be used directly for input into `contour()`, `filled.contour()`,
`image()` and others (see `lattice` package).
}
\description{
For now, this function only allows a Gaussian kernel for 2D density estimates.
This is probably most useful in making figures describing data with too many
coordinates to put in a single scatter diagram (`graphics::points()`).
}
\details{
?
}
\examples{
x <- c(rnorm(n=100, mean=5), rnorm(n=200, mean=3, sd=2))
y <- c(rnorm(n=100, mean=2), rnorm(n=200, mean=6, sd=2))
dens2d <- density2D(x=x, y=y, bw=0.5)
contour(x=dens2d$x, y=dens2d$y, z=dens2d$z)
}
