% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{etaSquaredTtest}
\alias{etaSquaredTtest}
\title{Get an eta-squared (for a t-test).}
\usage{
etaSquaredTtest(g1, g2, na.rm = TRUE)
}
\arguments{
\item{g1}{A numeric vector (scores for group 1).}

\item{g2}{A numeric vector (scores for group 2).}

\item{na.rm}{Boolean: remove NA's from g1 and g2.}
}
\value{
A number: the eta squared (effect size).
}
\description{
Calculate or bootstrap a confidence interval.
}
\details{
Eta-squared is a measure of effect size that expresses the 
propportion of variance explained by an effect. For a t-test, that effect
is the separation of the data into two groups. (This test does not work for 
t-testsof one group against a single value and does not differentiate 
between two-sample and paired-sample t-tests.)
}
\examples{
show example?
}
