% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{ezANOVA.pes}
\alias{ezANOVA.pes}
\title{ezANOVA wrapper to calculate partial eta squared}
\usage{
ezANOVA.pes(ezANOVAResult)
}
\arguments{
\item{ezANOVAResult}{ezANOVA result}
}
\value{
ezANOVA with partial eta-squared
}
\description{
ezANOVA wrapper to calculate partial eta squared
}
\details{
ezANOVA must be run with parameter 'detailed=T'.
}
\author{
Frank Papenmeier
}
