% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalFit.R, R/skewNormalFit.R
\name{fitNormal}
\alias{fitNormal}
\title{Fit a normal function to a set of values at given positions.}
\usage{
fitNormal(data)

fitNormal(data)
}
\arguments{
\item{data}{A numeric vector with columns x and y.}

\item{values}{A numeric vector of data values.}

\item{pos}{A numeric vector with N data positions.}

\item{output}{The kind of output wanted (see Value).}
}
\value{
By default, this function only returns the location of the
peak of the normal function fitted to the data, but there are other
options:
- mu: the mean
- sigma: the standard deviation
- offset: and offset added to the function
- scale: a largely irrelevant scale parameter
- par: all four parameters as a named, numeric vector

This function returns five parameters describing a skewed normal
function:
- mu: the mean
- sigma: the standard deviation
- lambda: the skew 
- scale: a largely irrelevant scale parameter
- offset: and offset added to the function
}
\description{
?

?
}
\details{
?

?
}
\examples{
?
?
}
