% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewNormalFit.R
\name{fitSkewNormal}
\alias{fitSkewNormal}
\title{Fit a skewed normal function to x and y coordinates.}
\usage{
fitSkewNormal(data)
}
\arguments{
\item{data}{A numeric vector with columns x and y.}
}
\value{
This function returns five parameters describing a skewed normal
function:
- mu: the mean
- sigma: the standard deviation
- lambda: the skew 
- scale: a largely irrelevant scale parameter
- offset: and offset added to the function
}
\description{
?
}
\details{
?
}
\examples{
?
}
