% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomID.R
\name{generateRandomIDs}
\alias{generateRandomIDs}
\title{Generate random participant IDs}
\usage{
generateRandomIDs(
  checkFile = "randomIDs.csv",
  IDlength = 3,
  number = 100,
  addToFile = checkFile
)
}
\arguments{
\item{checkFile}{CSV file with previous IDs that should not be duplicated}

\item{IDlength}{Number character pairs to use for IDs (3 is 6 characters)}

\item{number}{How many IDs to add to the existing list.}

\item{addToFile}{What CSV file to write the IDs to.}
}
\description{
This function generates a set of random IDs for use in 
experiments where participant anonymity is important. The IDs are not
generated with R's random number generator, but uses openssl.
}
\details{
The function does not yet check if it is possible to generate
as many unique IDs as requested with the given length, so be careful not to
start an endless loop.
}
\examples{
?
}
