% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reachProcessing.R
\name{getSplinedTrajectory}
\alias{getSplinedTrajectory}
\title{Get smooth splined coordinate interpolation over time}
\usage{
getSplinedTrajectory(x, y, t, length.out = length(t), spar = 0.01)
}
\arguments{
\item{x}{X-coordinates of a trajectory}

\item{y}{Y-coordinates of a trajectory}

\item{t}{Timestamps of the X and Y coordinates}

\item{spar}{Smoothing parameter for the spline, default: 0.01 (0.00-1.00)}
}
\value{
This function returns a data frame with smooth spline interpolated
trajectory given by x, y and t.
}
\description{
(Should one day allow data frame or matrix input as well...)
}
\details{
?
}
\examples{
?
}
