% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ginv.R
\name{ginv}
\alias{ginv}
\title{MASS ginv: generalized Moore-Penrose pseudo-inverse}
\usage{
ginv(X, tol = .Machine$double.eps * max(dim(X)))
}
\arguments{
\item{X}{A matrix.}

\item{tol}{Tolerance.}
}
\value{
Generalized Moore-Penrose pseudo inverse of matrix X.
}
\description{
This function is simply copied from the MASS package. The only
reason to do so is to reduce dependencies.
}
\details{
Not yet.
}
\examples{
library('SMCL')
M <- matrix(c(1:9),3,3)
Mi <- ginv(M)
zapsmall( M \%*\% Mi \%*\% M ) == zapsmall( M )
zapsmall( Mi \%*\% M \%*\% Mi ) == zapsmall( Mi )
zapsmall( t(M \%*\% Mi) ) == zapsmall( Mi \%*\% M )
zapsmall( t(Mi \%*\% M) ) == zapsmall( M \%*\% Mi )
}
