% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist2d.R
\name{hist2d}
\alias{hist2d}
\title{Get frequencies for a 2D histogram}
\usage{
hist2d(x, y = NA, nbins = c(25, 25), edges = NA)
}
\arguments{
\item{nbins}{A numeric vector with two integers: the number of bins for X 
and for Y coordinates, defaults to c(25, 25).}

\item{edges}{A list with the edges for the X and Y coordinates respectively.
This allows irregular bin sizes. Note that the number of edges is the number
of bins plus 1. If `edges` is given it overrides `nbins`.}

\item{X}{Vector of X-coordinates, or a Mx2 matrix with M sets of X and Y 
coordinates, or data frame with x and y column.}

\item{Y}{If X is not a matrix or data frame, but a vector this has to be a 
vector of equal length with Y coordinates.}
}
\value{
A list with three entries:
- `freq2D` numeric matrix with counts for each bin
- `x.edges` numeric vector with the edges used for x coordinates
- `y.edges` numeric vector with the edges used for y coordinates
}
\description{
?
}
\details{
?
}
\examples{
?
}
