% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{linPal}
\alias{linPal}
\title{Create linear palette in RGB color space}
\usage{
linPal(from = "#FFFFFF", to = "#E51636", n = 256, alpha = NULL)
}
\arguments{
\item{from}{first RGB color}

\item{to}{last RGB color}

\item{n}{number colors in output}

\item{alpha}{optional: if not NULL, alpha (transparency) value for all colors}
}
\value{
A palette of colors of length `n` where the first and last color are
equal to `from` and `to`, and the R, G and B values are linearly interpolated
between those in `from` and `to`.
}
\description{
Create a color gradient to use for plotting continuous values.
}
\examples{
?
}
