% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{localization_unaligned}
\alias{localization_unaligned}
\title{Non-aligned localization data points to illustrate circle-fitting}
\format{
Data frame with 25 rows and 8 columns:
\describe{
  \item{trial}{trial, numbered 0-24}
  \item{targetangle_deg}{angle of actual hand position relative to home}
  \item{homex_m}{home position x-coordinate in meters}
  \item{homey_m}{home position y-coordinate in meters}
  \item{handx_m}{hand position x-coordinate in meters}
  \item{handy_m}{hand position y-coordinate in meters}
  \item{tapx_px}{localization x-coordinate in pixels}
  \item{tapy_px}{localization y-coordinate in pixels}
}
}
\usage{
localization_unaligned
}
\description{
load with: data(localization)
}
\keyword{datasets}
