% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{mixCol}
\alias{mixCol}
\title{Linearly mix two colors given some weights.}
\usage{
mixCol(a = "#41ffc9", b = "#1d7791", balance = c(1, 1))
}
\arguments{
\item{a}{an RGB color string}

\item{b}{an RGB color string}

\item{balance}{two weights, one for each color, default: c(1,1)}
}
\value{
An RGB color string.
}
\description{
Returns a weighed mixed average of the two input colors.
Adds an alpha channel if not already present.
}
\examples{
?
}
