% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalFit.R
\name{normal}
\alias{normal}
\title{Evaluates a normal function at specific locations.}
\usage{
normal(par, x)
}
\arguments{
\item{par}{A named, numeric vector describing a normal function by its mean
('mu'), standard deviation ('sigma'), an offset ('offset') and a
multiplication factor ('scale').}

\item{x}{A numeric vector with locations to evaluate the normal function at.}
}
\value{
A numeric vector with the value of the normal function described by
the parameters `par`, as evaluated at the positions in `x`.
}
\description{
?
}
\details{
?
}
\examples{
?
}
