% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarHeatMap.R
\name{polarHeatMap}
\alias{polarHeatMap}
\title{Create a plot with a polar heatmap.}
\usage{
polarHeatMap(x, y, z, mincol = c(0.94, 0.98, 0.99), maxcol = c(0.06,
  0.82, 0.88), xlim = NA, ylim = NA, xunit = "degrees",
  border = NA, bordercol = "white", resolution = (1/180) * pi,
  alpha = 1)
}
\arguments{
\item{x}{Vector of M+1 angular bin edges.}

\item{y}{Vector of N+1 distance bin edges.}

\item{z}{Matrix (MxN) with two-dimensional histogram (frequency data), for 
example from SMCL::hist2d.}

\item{mincol}{Vector of length 3 with RGB values [0-1] for the minimum value
in z.}

\item{maxcol}{Vector of length 3 with RGB values [0-1] for the maximum value
in z. Other values in z are scaled, and each of the R, G and B values is 
linearly interpolated. No support for colormaps (yet).}

\item{xlim}{Vector with values limiting the x axis of the resulting plot.}

\item{ylim}{Vector with values limiting the y axis of the resulting plot. By
default, the is set to c(-1,1) for both x and y, as this will include the 
full polar figure. Only change these values when the data do not cover a
circular data set.}

\item{xunit}{The angular unit (of the x-coordinates). Can be 'degrees' or
'radians'. Will be converted to 'radians' if set in degrees.}

\item{border}{Draw borders around the parts of the heatmap. By default this
is set to NA, which means no border is drawn. If set to a numeric value, it
is used as the linewidth of the borders.}

\item{bordercol}{The color of the borders drawn around parts of the heatmap.
Default is 'white', can be any color accepted by base R graphics.}

\item{resolution}{The cicrular parts of the heatmap are approximated as a 
polygon, with at maximum this number of radians between vectors. Set to ~1
degree by default.}

\item{alpha}{The alpha value of the heatmap. Does not apply to the border.}
}
\value{
?
}
\description{
?
}
\details{
?
}
\examples{
?
}
