% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotateCoordinates.R
\name{rotateCoordinates}
\alias{rotateCoordinates}
\title{Rotate 2D trajectory}
\usage{
rotateCoordinates(df, angle, origin = c(0, 0))
}
\arguments{
\item{df}{A dataframe or matrix with two columns: X and Y coordinates.}

\item{angle}{An angle in degrees to rotate the trajectory by.}

\item{origin}{A vector with the coordinates to rotate around. Default (0,0)}
}
\value{
Data frame with the rotated trajectory coordinates.
}
\description{
Rotate a trajectory of X,Y coordinates.
}
\details{
Not yet.
}
\examples{
Not yet.
}
