% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{satCol}
\alias{satCol}
\title{Modify a color's saturation}
\usage{
satCol(col = "#41ffc9", sat.mult = 1.25)
}
\arguments{
\item{col}{an RGB color}

\item{sat.mult}{saturation multiplier}
}
\value{
An RGB color string.
}
\description{
Converts the RGB color to values in HSV space, then multiplies
the saturation component (S) by the given multiplier, leaving the hue (H)
and value (V) components as they were. The results is converted back to 
RGB strings before they are returned.
}
\examples{
?
}
