% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoRateModel.R
\name{seriesEffectiveSampleSize}
\alias{seriesEffectiveSampleSize}
\title{Estimate number of independent samples in a time series.}
\usage{
seriesEffectiveSampleSize(series, method = "ac_one")
}
\arguments{
\item{series}{numeric vector with the time series data.}

\item{method}{character specifying the method to find the number of 
independent samples in the time series data vector.

"ac_one": this uses the auto-correlation at lag 1 as ρ, and then gets
n_eff = n * ( (1−ρ)/(1+ρ) )
where n is the number of trials in the sequence (default)

"ac_lag.10": this divides the number of trials by the lag at which the auto-
correlation is about to go below 0.10:
n_eff = n / lag
(if AC is higher than 0.10, increase lag and redo, else use previous n_eff)

"ac_lag95%CI": this is similar to "ac_lag.10" but divides the number of 
trials by the lag at which the autocorrelation is about to be within the 95%
confidence interval, as determined by bootstrapping (with 1000 re-samples)}
}
\value{
numeric value with the estimated number of independent observation
}
\description{
This function is part of a set of functions to fit and 
evaluate the two-rate model of motor learning.
}
\details{
?
}
\examples{
?
}
