% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleLearningRates.R
\name{simpleRateModel}
\alias{simpleRateModel}
\title{Evaluate a simple learning rate model given parameters and a reach 
deviation schedule.}
\usage{
simpleRateModel(par, schedule)
}
\arguments{
\item{par}{A named vector with the model parameter (see details).}

\item{schedule}{A vector of length N with the perturbation schedule.}
}
\value{
A data frame with one column: `output`, and N rows, so that each row
has the output of the modeled process on each trials.
}
\description{
This function is part of a set of functions to fit and 
evaluate a simple learning rate model of motor learning.
}
\details{
The `par` argument is a named numeric vector that should have the
following element:
- LR: the learning rate
}
\examples{
?
}
