% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleLearningRates.R
\name{simpleRateMSE}
\alias{simpleRateMSE}
\title{Get the MSE for how well a simple learning rate model fits reaches.}
\usage{
simpleRateMSE(par, schedule, reaches)
}
\arguments{
\item{par}{A named numeric vector with the model parameter (see 
simpleRateModel).}

\item{schedule}{A numeric vector of length N with the perturbation schedule.}

\item{reaches}{A numeric vector of length N with reach deviations matching
the perturbation schedule.}
}
\value{
A float: the mean squared error between the total model output and
the reach deviations.
}
\description{
This function is part of a set of functions to fit and 
evaluate a simple learning rate model of motor learning.
}
\details{
The `par` argument is a named numeric vector that should have the
following element:
- LR: the learning rate
}
\examples{
?
}
