% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewNormalFit.R
\name{skewNormal}
\alias{skewNormal}
\title{Evaluates a normal function at specific locations.}
\usage{
skewNormal(par, x)
}
\arguments{
\item{par}{A named, numeric vector describing a skewed normal function by 
its mean ('mu'), standard deviation ('sigma'), skew ('lambda') an offset 
that is added ('offset') and a multiplication factor ('scale').}

\item{x}{A numeric vector with locations to evaluate the skewed normal 
function at.}
}
\value{
A numeric vector with the value of the skewed normal function 
described by the parameters `par`, as evaluated at the positions in `x`.
}
\description{
?
}
\details{
?
}
\examples{
?
}
