% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{t_col}
\alias{t_col}
\title{Set a color's alpha to a percentage.}
\usage{
t_col(color, percent = 50, name = NULL)
}
\arguments{
\item{color}{an RGB color}

\item{percent}{percentage transparency}

\item{name}{optional name for the color}
}
\value{
An RGB color string.
}
\description{
Sets the alpha value of the RGB color to match the given percentage
of transparency as closely as possible.
}
\examples{
?
}
