% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reachProcessing.R
\name{trimReach}
\alias{trimReach}
\title{Trim a data frame describing a reach to the interesting part.}
\usage{
trimReach(trialdf, homeStart = NA, targetReached = NA,
  velocity = 0.05, firstMove = FALSE, device = "hand",
  posunit = "pix", timeunit = "ms", homepos = c(0, 0))
}
\arguments{
\item{trialdf}{Data frame representing the reach.}

\item{homeStart}{If the participant has to get to the home position before
starting the out-and-back reach, this part could be trimmed. Set `homeStart`
to a numeric value expressing how close the `device` has to be to the home
position, specified in the same unit as `posunit`. The start of the trial
will be trimmed (not returned) up to when the device is that close to the 
start/home position. By default this is NA, so that this part is not 
trimmed.}

\item{targetReached}{If the return movement is represented in the data, this
may have to be trimmed as well. This parameter sets the distance at which 
the target is considered reached, and data after this point is trimmed.
The target position should be in columns named "targetx_[posunit]" and 
"targety_[posunit]".}

\item{velocity}{Very slow movement is usually not diagnostic. _After_ the
other parts of the data are trimmed, the instantaneous velocity is used as
a cut-off criterion: the first part of the reach that is under the velocity
criterion as a fraction of the maximum velocity in the whole reach, is 
trimmed. And either the final part that is below the velocity criterion is
trimmed, or everything after the first dip below the velocity criterion, 
depending on the `firstMove` parameter. Set to `NA` for no velocity 
criterion. By default this is set conservatively to 0.05.}

\item{firstMove}{Only used if the `velocity` parameter is not `NA`. If set 
to TRUE, the first part of the trajectory up to where it dips below the 
velocity criterion is kept (the rest is trimmed). If FALSE, only the final
part of the trajectory that goes below the velocity criterion is trimmed.}

\item{device}{The position columns to use are given by "[device]x_[posunit]"
in the `trialdf`, and similar for y. Can be something like 'hand', 'cursor',
'mouse', 'stylus' or 'robot'.}

\item{posunit}{The unit used for the x and y position data. Could 'pix' or
'cm', or whatever is used in the data.}

\item{timeunit}{The unit used for the time stamps of each sample. The column
names is "time_[timeunit]".}

\item{homepos}{The coordinates of the home position. Default is (0,0).}
}
\value{
Data frame describing the reach, minus the trimmed parts.
}
\description{
?
}
\details{
?
}
\examples{
?
}
