% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoRateModel.R
\name{twoRateFit}
\alias{twoRateFit}
\title{Fit the two-rate model to reach deviations.}
\usage{
twoRateFit(schedule, reaches, gridpoints = 6, gridfits = 6)
}
\arguments{
\item{schedule}{A vector of length N with the perturbation schedule.}

\item{reaches}{A vector of length N with reach deviation data.}
}
\value{
A named numeric vector with the optimal parameters that fit the two
rate model to the data as best as possible, with these elements:
- Ls: the slow learning rate
- Lf: the fast learning rate
- Rs: the slow retention rate
- Rf: the fast retention rate
}
\description{
This function is part of a set of functions to fit and 
evaluate the two-rate model of motor learning.
}
\details{
?
}
\examples{
data("localization")

baseline <- function(reachvector,blidx) reachvector - mean(reachvector[blidx], na.rm=TRUE)
tworatedata[,4:ncol(tworatedata)] <- apply(tworatedata[,4:ncol(tworatedata)], FUN=baseline, MARGIN=c(2), blidx=c(17:32))
reaches <- apply(tworatedata[4:ncol(tworatedata)], FUN=median, MARGIN=c(1), na.rm=TRUE)
schedule <- tworatedata$schedule
towRateFit(schedule, reaches)
}
