% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoRateModel.R
\name{twoRates}
\alias{twoRates}
\title{Evaluate two-rate model fits.}
\usage{
twoRates(groupdata, estN = "ac_one", compOne = FALSE)
}
\arguments{
\item{groupdata}{A named vector with data for each group. Each element
should be a list (or data frame) with 2 named entries (or columns):

"schedule": the perturbation for each trial, used for fitting models

"reaches": the actual reaches made by this group on each trial, also used
for fitting models (if estN='aclag' reaches can't contain missing values)

The labels in the vector should correspond to group names. They will be used
in the output, and to match specifications of N to the correct data.}

\item{estN}{This is either a string, specifying a method to estimate the
number of observations for calculating an AIC. Alternatively, this is a 
vector with named entries where the names are equal to the groups in 
`groupdata` and the value is the N to use for that group.

Named methods are the same as in `seriesEffectiveSampleSize`.}

\item{compOne}{A boolean specifying whether or not to compare the two-rate
model fit to a one-rate model fit for each group.}
}
\value{
A data frame with a row for each group. This will give the estimated
parameter values, an MSE, N and AIC. It will list relative log likelihoods 
too, comparing the groups two-rate AICs, as well as within each group the 
AICs for a one- and two-rate model fit, if requested (see `compOne`).
}
\description{
This function is part of a set of functions to fit and 
evaluate the two-rate model of motor learning.
}
\details{
?
}
\examples{
?
}
