% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zerocrossings.R
\name{zerocrossings}
\alias{zerocrossings}
\title{Where does a signal cross zero?}
\usage{
zerocrossings(v, bounces = FALSE)
}
\arguments{
\item{v}{A numeric vector containing the signal.}

\item{bounces}{A boolean: do bounces (a zero without sign change) count?}
}
\value{
A vector with indices into `v` where the signal will cross zero.
}
\description{
This is a first attempt at a generic function to find zero
crossings in arbitrary signals. The indices that are returned, are the 
samples right before the signal crosses zero, and if bounces are allowed,
also the first sample of every bounce. Supposedly.
}
\examples{
?
}
