% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalFit.R
\name{fitNormal}
\alias{fitNormal}
\title{Fit a normal function to a set of values at given positions.}
\usage{
fitNormal(values, pos, output = "mu")
}
\arguments{
\item{values}{A numeric vector of data values.}

\item{pos}{A numeric vector with N data positions.}

\item{output}{The kind of output wanted (see Value).}
}
\value{
By default, this function only returns the location of the
peak of the normal function fitted to the data, but there are other
options:
- mu: the mean
- sigma: the standard deviation
- offset: and offset added to the function
- scale: a largely irrelevant scale parameter
- par: all four parameters as a named, numeric vector
}
\description{
?
}
\details{
?
}
\examples{
?
}
