% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReachAngleAt.R
\name{getReachAngleAt}
\alias{getReachAngleAt}
\title{Angular deviation of a reach from target at a specific point.}
\usage{
getReachAngleAt(trialdf, location = "maxvel")
}
\arguments{
\item{trialdf}{Data frame representing the reach.}

\item{location}{Specifies which location to use.}
}
\value{
Matrix with 1 row, 2 columns. First column has the angular deviation
of the reach from tha target angle at the specified location in degrees. The
second column has the angle of the reach in degrees.
}
\description{
?
}
\details{
?
}
\examples{
?
}
