% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSplinedVelocity.R
\name{getSplinedVelocity}
\alias{getSplinedVelocity}
\title{Get velocity profile after spline interpolation}
\usage{
getSplinedVelocity(x, y, t, spar = 0.01)
}
\arguments{
\item{t}{Timestamps of the X and Y coordinates}

\item{spar}{Smoothing parameter for the spline, default: 0.01 (0.00-1.00)}

\item{X}{X-coordinates of a trajectory}

\item{Y}{Y-coordinates of a trajectory}
}
\value{
This function returns a data frame with spline interpolated velocity
and time for the trajectory given by X, Y and t.
}
\description{
?
}
\details{
?
}
\examples{
?
}
