% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinv.R
\name{ginv}
\alias{ginv}
\title{MASS ginv: generalized Moore-Penrose pseudo-inverse}
\usage{
ginv(X, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{X}{A matrix.}

\item{tol}{Tolerance.}
}
\value{
Generalized Moore-Penrose pseudo inverse of matrix X.
}
\description{
This function is simply copied from the MASS package. The only
reason to do so is to reduce dependencies.
}
\details{
Not yet.
}
\examples{
Not yet.
}
